+++ CHANGES.md
@@ -68,6 +68,7 @@ Changes in 12.5
 - Module is addon, set custom module icon & remove $icon param in DashboardModule.fnc.php
 - Redirection is done in HTML in HackingLog.fnc.php
 - Fix reload theme when "Force" checked in Theme.fnc.php
+- Redirection is done in HTML in Theme.fnc.php
 
 Changes in 12.4.3
 -----------------